fix(portfolio): support fractional shares in pp-cli JAR
- Main.java: Long.parseLong → Double.parseDouble for shares
- PpClient.java: long shares → double shares (Values.Share.factorize(double) exists)
- build.sh: add Maven pp-cli build step (was in deploy.sh — wrong place)
- deploy.sh: remove pp-cli build (now in build.sh)
- java_bridge.js: remove Math.round(shares) (done in prior commit)

+++ modules/portfolio-tracker/pp-cli/src/main/java/name/abuchen/portfolio/cli/Main.java
@@ -41,7 +41,7 @@ public class Main {
                             params.get("security-id"),
                             require(params, "type"),
                             require(params, "date"),
-                            Long.parseLong(params.getOrDefault("shares", "0")),
+                            Double.parseDouble(params.getOrDefault("shares", "0")),
                             Double.parseDouble(require(params, "price")),
                             require(params, "currency"),
                             Double.parseDouble(params.getOrDefault("fees", "0")),
